fix(http): implement raw body and reseting headers
Json body doesn't exist in global variables but in php://input stream,
so for testing it requires to add $rawBody in HttpRouterTester, to be
able to mock that body.
The next change is the problem that I found during testing my change.
HttpRouterTester is setting up headers in global variables, so if some
headers has been set, they exists for every next test to the end of PHP
process