fix(http): implement json body parser for server request
To fix issue with parsing json body in Laminas requests, I've
implemented parser class. It's an invokable class that takes data from
body stream, parse it and add to request as parsedBody.

I've also considered implementing that as ServerRequestFilter for
Laminas package. As it's described it allows us to modify request, but
seems that it's created for filtering purpose, as name says.

The next possible solution could be implementing some middleware, but
unfortunately Tempest mddlewares afaik, are triggered only after the
controller.