Make directory matching platform agnostic
The vendor directory is string matched to give it the lowest priority when loading configuration files during discovery. This change sets the prefixed and suffixed directory separator to the PHP constant DIRECTORY_SEPARATOR rather than hard-coding '/'. This allows the match to work when PHP is running on Windows - which uses '\' as a directory separator - as well as on other platforms.