Mark password as sensitive parameter to avoid accidental exposure.
User passwords should not end up in stack traces or logs.

Most important is adding it to the UserModel stub that is used by the `tempest install auth`.

In the documetation for Hashed this example also shows up, but since it would distract from the purpose of the example there, I've chosen to omit it.